[Epic-ID: ODUHIGH-516][Task-ID: ODUHIGH-529] RIC Indication

+++ src/du_app/du_tmr.h
@@ -21,7 +21,7 @@
 
 short int duActvTmr(Ent ent,Inst inst); 
 bool duChkTmr(PTR cb, int16_t tmrEvnt);
-void duStartTmr(PTR cb, int16_t tmrEvnt, uint8_t timerValue);
+void duStartTmr(PTR cb, int16_t tmrEvnt, uint32_t timerValue);
 
 /**********************************************************************
   End of file
